requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
paho-mqtt==2.1.0
cbor2==5.6.0
asyncpg==0.29.0
python-dotenv==1.0.1
